+++ src/app/data-evaluation/facility/analysis/run-analysis/group-analysis/regression-model-selection/regression-model-selection.component.css
@@ -38,3 +38,67 @@ tr.selected-model{
   background-color: #e5e8ea;
   color: red;
 }
+
+.dropdown {
+    position: relative;
+    display: inline-block;
+}
+
+.dropdown-btn {
+    padding: 8px 12px;
+    cursor: pointer;
+    border-radius: 6px;
+    border: 1px solid #0078d4;
+    background: #f5faff;
+    color: #0078d4;
+    font-weight: 600;
+}
+
+.dropdown-btn:hover {
+    background: #e6f0fa;
+}
+
+.alert-style {
+    display: flex;
+    align-items: center;
+    vertical-align: middle;
+}
+
+.dropdown-content {
+    display: block;
+    position: absolute;
+    background-color: #fff;
+    border: 1px solid #0078d4;
+    border-radius: 8px;
+    max-height: 240px;
+    overflow-y: auto;
+    padding: 8px;
+    width: 150px;
+    z-index: 10;
+    margin-top: 4px;
+    right: 0;
+    left: auto;
+}
+
+.multiselect-option {
+    display: flex;
+    align-items: center;
+    padding: 8px 0;
+    font-size: 1.08rem;
+    border-bottom: 1px solid #f0f0f0;
+}
+.multiselect-option:last-child {
+    border-bottom: none;
+}
+
+.multiselect-option:hover {
+    background: #f5faff;
+}
+
+.multiselect-option input[type="checkbox"] {
+    width: 18px;
+    height: 18px;
+    margin-right: 12px;
+    border-radius: 4px;
+    border: 2px solid #0078d4;
+}
